+++ src/Utility/FileUtility.php
@@ -2,6 +2,10 @@
 
 namespace SourceBroker\DeployerLoader\Utility;
 
+use RecursiveDirectoryIterator;
+use RecursiveIteratorIterator;
+use SplFileInfo;
+
 /**
  * Class FileUtility
  *
@@ -16,18 +20,18 @@ class FileUtility
     public function requireFilesFromDirectoryReqursively($absolutePath, $excludePattern = null)
     {
         if (is_dir($absolutePath)) {
-            $iterator = new \RecursiveIteratorIterator(
-                new \RecursiveDirectoryIterator($absolutePath),
-                \RecursiveIteratorIterator::SELF_FIRST
+            $iterator = new RecursiveIteratorIterator(
+                new RecursiveDirectoryIterator($absolutePath),
+                RecursiveIteratorIterator::SELF_FIRST
             );
             foreach ($iterator as $file) {
-                /** @var $file \SplFileInfo */
+                /** @var $file SplFileInfo */
                 if ($file->isFile()) {
                     $excludeMatch = null;
                     if ($excludePattern !== null) {
                         $excludeMatch = preg_match($excludePattern, $file->getFilename());
                     }
-                    if ($file->getExtension() == 'php' && $excludeMatch !== 1) {
+                    if ($excludeMatch !== 1 && $file->getExtension() === 'php') {
                         /** @noinspection PhpIncludeInspection */
                         require_once $file->getRealPath();
                     }