+++ CHANGELOG.rst
@@ -1,6 +1,12 @@
 Changelog
 ---------
 
+master
+~~~~~~
+
+1) [BUGFIX] Fix wrong exception with multiple repositories with same namespace prefix (sourcebroker/deployer-loader/pull/1)
+2) [TASK] Refactor autoload.php and FileUtility / Load classes.
+
 2.0.0
 ~~~~~