+++ CHANGELOG.rst
@@ -6,6 +6,7 @@ master
 
 1) [BUGFIX] Fix wrong exception with multiple repositories with same namespace prefix (sourcebroker/deployer-loader/pull/1)
 2) [TASK] Refactor autoload.php and FileUtility / Load classes.
+3) [TASK] Refactor docs.
 
 2.0.0
 ~~~~~