+++ CHANGELOG.rst
@@ -4,6 +4,12 @@ Changelog
 master
 ~~~~~~
 
+1) [TASK] Simplify autoload.php with require instead of include.
+
+
+2.0.1
+~~~~~
+
 1) [BUGFIX] Fix wrong exception with multiple repositories with same namespace prefix (sourcebroker/deployer-loader/pull/1)
 2) [TASK] Refactor autoload.php and FileUtility / Load classes.
 3) [TASK] Refactor docs.